pointerevents/pointerevent_pointerOver.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: Dispatch pointerover. </title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <meta name="assert" content="When a pointing device is moved into the hit test boundaries of an element, the pointerover event must be dispatched. "/>

        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
		<script type="text/javascript">

        setup({ explicit_done: true });

        var detected_pointertypes = {};
        add_completion_callback(showPointerTypes);

        var target0;;
        var test_pointerOver;

        function run() {
            target0 = document.getElementById("target0");
            test_pointerOver = async_test("pointerover is dispatched");  // set up test harness

            // set pointer capture
            on_event(target0, "pointerover", function (event) {
                
                test_pointerOver.step(function () {
                    check_PointerEvent(event);
                    assert_true(event.type == "pointerover", "pointerover received");
                });

                detected_pointertypes[event.pointerType] = true;
                test_pointerOver.done(); // complete tes
            });
        }

        function showPointerTypes() {
            var complete_notice = document.getElementById("complete-notice");
            var pointertype_log = document.getElementById("pointertype-log");
            var pointertypes = Object.keys(detected_pointertypes); pointertype_log.innerHTML = pointertypes.length ? pointertypes.join(",") : "(none)";
            if (test_pointerOver.status == test_pointerOver.NOTRUN
                || test_pointerOver.status == test_pointerOver.TIMEOUT) {
                complete_notice.innerHTML = "Test FAIL due to time out.  This is due to the browser not supporting pointer, or the manual steps were not run to completion.";
                complete_notice.style.background = "red";
            }
            complete_notice.style.display = "block";
        }
		</script>

		<style>
			div {
				margin: 0em;
				padding: 2em;
			}

			#target0 {
				touch-action: none;
				background: yellow;
				border: 1px solid orange;
			}

			#complete-notice {
				border: 1px solid #0a0;
				display: none;
			}

			#pointertype-log {
				font-weight: bold;
			}
		</style>
	</head>
	<body onload="run()">
		<h1>Pointer Event: Dispatch pointerover.</h1>
		<h4>Test Description: 
				When a pointing device is moved into the hit test boundaries of an element, the pointerover event must be dispatched.  
		</h4>
		<br />
		<div id="target0">
			Use mouse, touch or pen to hover or contact this element..
		</div>
		<div id="complete-notice">
			<p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
			<p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
			<p>Refresh the page to run the tests again with a different pointer type?</p>
		</div>

		<div id="log"></div>
	</body>
</html>
